mean field of 3 nearest neighbors

+++ rainbow/train.py
@@ -250,14 +250,14 @@ def main(args):
                             complements = torch.tensor([0]).repeat(1, (4-N))[0]
                             neighbor_actions = torch.cat([neighbor_actions, complements])
                             neighbor_actions = torch.nn.functional.one_hot(neighbor_actions, num_classes=2)
-                            neighbor_actions = torch.mean(neighbor_actions, dim=0)
+                            neighbor_actions = torch.mean(neighbor_actions.float(), dim=0)
                             mean_fields[j] = neighbor_actions
                     else:
                         for j in range(N):
                             neighbors = np.argsort(distance_matrix[j])[1:4]
                             neighbor_actions = actions_[neighbors]
                             neighbor_actions = torch.nn.functional.one_hot(neighbor_actions, num_classes=2)
-                            neighbor_actions = torch.mean(neighbor_actions, dim=0)
+                            neighbor_actions = torch.mean(neighbor_actions.float(), dim=0)
                             mean_fields[j] = neighbor_actions
 
                     for j in range(N):
